e2e: decrease progresstimeout to 1m
Shipper is fast enough that it can finish its work in under 1 minute
now. This makes failures faster, as it'll now take half the time for e2e
to report timeouts. This is great when you know you have a flaky test
that you want to go green so you can actually merge important changes
^_^