e2e tests: skip target steps when testing aborts
Shipper doesn't care about following strategies step by step: it'll
converge to whatever you set. Since we're testing aborts, and not
strategy progress, there's no need to wait for each step (which just
consumes time), so we'll just skip directly to the step we want the
release to be in.