release controller: check InstallationTarget Conditions
... instead of Cluster Status. We're moving towards relocating the smarts in the
checks for a target's readiness away from the executor and into the
controllers themselves, leaving the executor to check for a single
`Ready` condition at the target's top level status. This commit is not
yet that, the first step is to stop looking into Status fields, and
eventually drop them. InstallationTarget is the only one that currently
gets its Cluster Statuses checked, and now that's been changed into a
check for Cluster `Ready` conditions.